+++ src/main/java/com/wenxin2/marioverse/blocks/entities/DisguisedBlockEntity.java
@@ -1,5 +1,6 @@
 package com.wenxin2.marioverse.blocks.entities;
 
+import com.wenxin2.marioverse.blocks.properties.BlockStatePropertyRegistry;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.BuiltInRegistries;
@@ -12,19 +13,21 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
+import net.neoforged.neoforge.client.model.data.ModelData;
+import net.neoforged.neoforge.client.model.data.ModelProperty;
 import net.neoforged.neoforge.items.IItemHandler;
 import net.neoforged.neoforge.items.ItemStackHandler;
+import org.jetbrains.annotations.NotNull;
 
-public class DisguiseBlockEntity extends BlockEntity {
+public class DisguisedBlockEntity extends BlockEntity {
+    public static final ModelProperty<BlockState> DISGUISED = new ModelProperty<>();
     protected BlockState disguiseState = Blocks.AIR.defaultBlockState();
 
     protected final ItemStackHandler inventory = new ItemStackHandler(1) {
         @Override
         protected void onContentsChanged(int slot) {
-            setChanged();
-
-            if (slot == 0 && level != null)
-                DisguiseBlockEntity.this.updateDisguise();
+            if (slot == 0 && DisguisedBlockEntity.this.level != null)
+                DisguisedBlockEntity.this.setDisguise();
         }
 
         @Override
@@ -38,19 +41,34 @@ public class DisguiseBlockEntity extends BlockEntity {
         }
     };
 
-    public DisguiseBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state) {
+    public DisguisedBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state) {
         super(type, pos, state);
     }
 
-    protected void updateDisguise() {
-        ItemStack stack = inventory.getStackInSlot(0);
+    protected void setDisguise() {
+        ItemStack stack = this.inventory.getStackInSlot(0);
+        boolean disguised = false;
 
-        if (stack.getItem() instanceof BlockItem blockItem)
+        if (stack.getItem() instanceof BlockItem blockItem) {
             this.disguiseState = blockItem.getBlock().defaultBlockState();
+            disguised = true;
+        }
         else this.disguiseState = Blocks.AIR.defaultBlockState();
 
-        if (level != null)
-            level.sendBlockUpdated(worldPosition, getBlockState(), getBlockState(), 3);
+        if (this.level != null) {
+            BlockState state = this.level.getBlockState(this.worldPosition);
+
+            if (state.hasProperty(BlockStatePropertyRegistry.DISGUISED)
+                    && state.getValue(BlockStatePropertyRegistry.DISGUISED) != disguised)
+                this.level.setBlock(this.worldPosition, state.setValue(BlockStatePropertyRegistry.DISGUISED, disguised), 3);
+            if (state.hasProperty(BlockStatePropertyRegistry.INVISIBLE)
+                    && state.hasProperty(BlockStatePropertyRegistry.DISGUISED)
+                    && state.getValue(BlockStatePropertyRegistry.DISGUISED))
+                this.level.setBlock(this.worldPosition, state.setValue(BlockStatePropertyRegistry.INVISIBLE, true), 3);
+            this.requestModelDataUpdate();
+            this.level.sendBlockUpdated(this.worldPosition, this.getBlockState(), this.getBlockState(), Block.UPDATE_CLIENTS);
+            this.setChanged();
+        }
     }
 
     public BlockState getDisguise() {
@@ -58,21 +76,21 @@ public class DisguiseBlockEntity extends BlockEntity {
     }
 
     public void setDisguiseItem(ItemStack stack) {
-        inventory.setStackInSlot(0, stack);
+        this.inventory.setStackInSlot(0, stack);
     }
 
     public ItemStack getDisguiseItem() {
-        return inventory.getStackInSlot(0);
+        return this.inventory.getStackInSlot(0);
     }
 
     public IItemHandler getInventory() {
-        return inventory;
+        return this.inventory;
     }
 
     @Override
     protected void saveAdditional(CompoundTag tag, HolderLookup.Provider provider) {
         super.saveAdditional(tag, provider);
-        tag.put("inventory", inventory.serializeNBT(provider));
+        tag.put("inventory", this.inventory.serializeNBT(provider));
 
         if (!this.disguiseState.isAir())
             tag.putString("disguiseBlock", BuiltInRegistries.BLOCK.getKey(this.disguiseState.getBlock()).toString());
@@ -81,11 +99,17 @@ public class DisguiseBlockEntity extends BlockEntity {
     @Override
     protected void loadAdditional(CompoundTag tag, HolderLookup.Provider provider) {
         super.loadAdditional(tag, provider);
-        inventory.deserializeNBT(provider, tag.getCompound("inventory"));
+        this.inventory.deserializeNBT(provider, tag.getCompound("inventory"));
 
         if (tag.contains("disguiseBlock")) {
             Block block = BuiltInRegistries.BLOCK.get(ResourceLocation.parse(tag.getString("disguiseBlock")));
             this.disguiseState = block.defaultBlockState();
         }
+    }
+
+    @NotNull
+    @Override
+    public ModelData getModelData() {
+        return ModelData.builder().with(DISGUISED, this.getDisguise()).build();
     }
 }