+++ src/main/java/com/wenxin2/marioverse/blocks/properties/BlockStatePropertyRegistry.java
@@ -1,0 +1,8 @@
+package com.wenxin2.marioverse.blocks.properties;
+
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+
+public class BlockStatePropertyRegistry {
+    public static final BooleanProperty DISGUISED = BooleanProperty.create("disguised");
+    public static final BooleanProperty INVISIBLE = BooleanProperty.create("invisible");
+}