+++ src/main/java/com/wenxin2/marioverse/client/models/blocks/DisguisedBlockBakedModel.java
@@ -1,0 +1,77 @@
+package com.wenxin2.marioverse.client.models.blocks;
+
+import com.wenxin2.marioverse.Marioverse;
+import com.wenxin2.marioverse.blocks.entities.BlockSpawnerBlockEntity;
+import java.util.Collections;
+import java.util.List;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.block.model.BakedQuad;
+import net.minecraft.client.renderer.block.model.ItemOverrides;
+import net.minecraft.client.renderer.texture.TextureAtlas;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.core.Direction;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.block.state.BlockState;
+import net.neoforged.neoforge.client.model.IDynamicBakedModel;
+import net.neoforged.neoforge.client.model.data.ModelData;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class DisguisedBlockBakedModel implements IDynamicBakedModel {
+    @NotNull
+    @Override
+    public List<BakedQuad> getQuads(@Nullable BlockState state, @Nullable Direction side, RandomSource random) {
+        return getQuads(state, side, random, ModelData.EMPTY, null);
+    }
+
+    @NotNull
+    @Override
+    public List<BakedQuad> getQuads(@Nullable BlockState state, @Nullable Direction side, RandomSource random,
+            ModelData data, @Nullable RenderType renderType) {
+        BlockState disguise = data.get(BlockSpawnerBlockEntity.DISGUISED);
+
+        if (disguise == null || disguise.isAir())
+            return Collections.emptyList();
+
+        BakedModel model = Minecraft.getInstance().getBlockRenderer().getBlockModel(disguise);
+
+        return model.getQuads(disguise, side, random, data, renderType);
+    }
+
+    @NotNull
+    @Override
+    public ItemOverrides getOverrides() {
+        return ItemOverrides.EMPTY;
+    }
+
+    @Override
+    public boolean useAmbientOcclusion() {
+        return true;
+    }
+
+    @Override
+    public boolean isGui3d() {
+        return true;
+    }
+
+    @Override
+    public boolean usesBlockLight() {
+        return true;
+    }
+
+    @Override
+    public boolean isCustomRenderer() {
+        return false;
+    }
+
+    @NotNull
+    @Override
+    public TextureAtlasSprite getParticleIcon() {
+        return Minecraft.getInstance()
+                .getTextureAtlas(TextureAtlas.LOCATION_BLOCKS)
+                .apply(ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "block/block_spawner"));
+    }
+}