+++ changelog.md
@@ -1,6 +1,10 @@
 ## Update 1.8.3
 - Add Block Spawner
-- Fix shift-clicking into slots of the ? Block GUI
+  - Right-click to open gui
+  - Add a block to the top slot for replacing in world
+  - Add a block to the bottom slot to disguise the block spawner
+- Add slot icon to refill GUI
+- Fix shift-clicking into slots of the refill GUI
 
 ## Update 1.8.2
 - Add dispenser behavior for equipping carrots on Show Pokeys